apps/account_manager/seralizers/bank_serializer.py
from rest_framework import serializers

from apps.account_manager.models import Bank


class BankSerializer(serializers.ModelSerializer):
    class Meta:
        model = Bank
        fields = "__all__"
